requirements.txt
django
djangorestframework
python-dotenv
pandas
mysqlclient